+++ test.py
@@ -1,12 +1,14 @@
-from contextlib import contextmanager
 import logging
 import os
 import re
 import stat
 import sys
 import unittest
-from os.path import expanduser, join, abspath
+from contextlib import contextmanager
+from os.path import abspath, expanduser, join
 from textwrap import dedent
+
+from config_resolver import Config, NoVersionError, SecuredConfig, get_config
 
 try:
     from ConfigParser import NoOptionError, NoSectionError
@@ -23,11 +25,6 @@ except ImportError:
     except ImportError:
         have_mock = False
 
-from config_resolver import (
-    Config,
-    SecuredConfig,
-    NoVersionError,
-)
 
 
 @contextmanager
@@ -72,11 +69,13 @@ class TestableHandler(logging.Handler):
     def assert_contains(self, logger, level, needle):
         if not self.contains(logger, level, needle):
             msg = '%s did not contain a message with %r and level %r'
+            level = logging.getLevelName(level)
             raise AssertionError(msg % (logger, needle, level))
 
     def assert_contains_regex(self, logger, level, needle):
         if not self.contains(logger, level, needle, is_regex=True):
             msg = '%s did not contain a message matching %r and level %r'
+            level = logging.getLevelName(level)
             raise AssertionError(msg % (logger, needle, level))
 
     def contains(self, logger, level, message, is_regex=False):
@@ -102,13 +101,34 @@ class TestableHandler(logging.Handler):
         return False
 
 
+class TestBase(unittest.TestCase):
+    '''
+    A superclass for tests to make use of the TestableHandler class in each
+    test.
+
+    This would be easier using the pytest caplog fixture, but this would entail
+    too many code changes and conflicts down the line.
+    '''
+
+    def setUp(self):
+        logger = logging.getLogger('config_resolver')
+        logger.setLevel(logging.DEBUG)
+        self.catcher = TestableHandler()
+        logger.addHandler(self.catcher)
+        def remove_handler():
+            logger.removeHandler(self.catcher)
+        self.addCleanup(remove_handler)
+
+
+
 @unittest.skipUnless(sys.version_info > (3, 0), 'Test only valid in Python 2')
-class SimpleInitFromContent(unittest.TestCase):
+class SimpleInitFromContent(TestBase):
     '''
     Tests loading a config string from memory
     '''
 
     def setUp(self):
+        super(SimpleInitFromContent, self).setUp()
         self.cfg = Config('not', 'existing', search_path='testdata')
         self.cfg.read_string(dedent(
             '''\
@@ -124,9 +144,10 @@ class SimpleInitFromContent(unittest.TestCase):
         self.assertEqual(self.cfg.get('section_mem', 'val'), '1')
 
 
-class SimpleInitTest(unittest.TestCase):
+class SimpleInitTest(TestBase):
 
     def setUp(self):
+        super(SimpleInitTest, self).setUp()
         self.cfg = Config('hello', 'world', search_path='testdata')
 
     def test_simple_init(self):
@@ -144,7 +165,7 @@ class SimpleInitTest(unittest.TestCase):
         self.assertIs(self.cfg.get('a', 'b', default=None), None)
 
 
-class AdvancedInitTest(unittest.TestCase):
+class AdvancedInitTest(TestBase):
 
     def tearDown(self):
         os.environ.pop('HELLO_WORLD_PATH', None)
@@ -166,14 +187,10 @@ class AdvancedInitTest(unittest.TestCase):
 
     def test_env_name_override(self):
         os.environ['HELLO_WORLD_FILENAME'] = 'test.ini'
-        logger = logging.getLogger('config_resolver')
-        logger.setLevel(logging.DEBUG)
-        catcher = TestableHandler()
-        logger.addHandler(catcher)
         Config('hello', 'world')
         msg = ("filename was overridden with 'test.ini' by the environment "
                "variable HELLO_WORLD_FILENAME")
-        catcher.assert_contains(
+        self.catcher.assert_contains(
             'config_resolver.hello.world',
             logging.INFO,
             msg)
@@ -189,15 +206,11 @@ class AdvancedInitTest(unittest.TestCase):
             expected)
 
     def test_env_path_override_log(self):
-        logger = logging.getLogger('config_resolver')
-        logger.setLevel(logging.DEBUG)
         os.environ['HELLO_WORLD_PATH'] = 'testdata:testdata/a:testdata/b'
-        catcher = TestableHandler()
-        logger.addHandler(catcher)
         Config('hello', 'world')
         msg = ("overridden with 'testdata:testdata/a:testdata/b' by the "
                "environment variable 'HELLO_WORLD_PATH'")
-        catcher.assert_contains(
+        self.catcher.assert_contains(
             'config_resolver.hello.world',
             logging.INFO,
             msg)
@@ -219,15 +232,11 @@ class AdvancedInitTest(unittest.TestCase):
             expected)
 
     def test_env_path_add_log(self):
-        logger = logging.getLogger('config_resolver')
-        logger.setLevel(logging.DEBUG)
         os.environ['HELLO_WORLD_PATH'] = '+testdata:testdata/a:testdata/b'
-        catcher = TestableHandler()
-        logger.addHandler(catcher)
         Config('hello', 'world')
         msg = ("extended with ['testdata', 'testdata/a', 'testdata/b'] by the "
                "environment variable HELLO_WORLD_PATH")
-        catcher.assert_contains(
+        self.catcher.assert_contains(
             'config_resolver.hello.world',
             logging.INFO,
             msg)
@@ -252,7 +261,7 @@ class AdvancedInitTest(unittest.TestCase):
         self.assertEqual(cfg.app_name, 'world')
 
 
-class FunctionalityTests(unittest.TestCase):
+class FunctionalityTests(TestBase):
 
     def test_mandatory_section(self):
         config = Config('hello', 'world', search_path='testdata')
@@ -265,16 +274,12 @@ class FunctionalityTests(unittest.TestCase):
             config.get('section1', 'nosuchoption')
 
     def test_unsecured_logmessage(self):
-        logger = logging.getLogger('config_resolver')
-        logger.setLevel(logging.DEBUG)
-        catcher = TestableHandler()
-        logger.addHandler(catcher)
         SecuredConfig('hello', 'world', filename='test.ini',
                       search_path='testdata')
         expected_message = (
             "File 'testdata/test.ini' is not secure enough. "
             "Change it's mode to 600")
-        catcher.assert_contains(
+        self.catcher.assert_contains(
             'config_resolver.hello.world',
             logging.WARNING,
             expected_message)
@@ -314,22 +319,17 @@ class FunctionalityTests(unittest.TestCase):
             Config('hello', 'world', search_path='testdata', version='1.1')
 
     def test_mismatching_major(self):
-        logger = logging.getLogger('config_resolver')
-        logger.setLevel(logging.DEBUG)
-        catcher = TestableHandler()
-        logger.addHandler(catcher)
-
         config = Config('hello', 'world', search_path='testdata/versioned',
                         version='1.1')
-        catcher.assert_contains(
+        self.catcher.assert_contains(
             'config_resolver.hello.world',
             logging.ERROR,
             'Invalid major version number')
-        catcher.assert_contains(
+        self.catcher.assert_contains(
             'config_resolver.hello.world',
             logging.ERROR,
             '2.1')
-        catcher.assert_contains(
+        self.catcher.assert_contains(
             'config_resolver.hello.world',
             logging.ERROR,
             '1.1')
@@ -342,21 +342,17 @@ class FunctionalityTests(unittest.TestCase):
         self.assertEqual(config.loaded_files, [])
 
     def test_mismatching_minor(self):
-        logger = logging.getLogger('config_resolver')
-        logger.setLevel(logging.DEBUG)
-        catcher = TestableHandler()
-        logger.addHandler(catcher)
         Config('hello', 'world', search_path='testdata/versioned',
                version='2.0')
-        catcher.assert_contains(
+        self.catcher.assert_contains(
             'config_resolver.hello.world',
             logging.WARNING,
             'Mismatching minor version number')
-        catcher.assert_contains(
+        self.catcher.assert_contains(
             'config_resolver.hello.world',
             logging.WARNING,
             '2.1')
-        catcher.assert_contains(
+        self.catcher.assert_contains(
             'config_resolver.hello.world',
             logging.WARNING,
             '2.0')
@@ -368,22 +364,18 @@ class FunctionalityTests(unittest.TestCase):
         files even if the application did not explicitly request a version
         number!
         """
-        logger = logging.getLogger('config_resolver')
-        logger.setLevel(logging.DEBUG)
-        catcher = TestableHandler()
-        logger.addHandler(catcher)
         Config('hello', 'world',
                filename='mismatch.ini',
                search_path='testdata/versioned:testdata/versioned2')
-        catcher.assert_contains(
+        self.catcher.assert_contains(
             'config_resolver.hello.world',
             logging.ERROR,
             'Invalid major version number')
-        catcher.assert_contains(
+        self.catcher.assert_contains(
             'config_resolver.hello.world',
             logging.ERROR,
             '1.0')
-        catcher.assert_contains(
+        self.catcher.assert_contains(
             'config_resolver.hello.world',
             logging.ERROR,
             '2.0')
@@ -458,12 +450,8 @@ class FunctionalityTests(unittest.TestCase):
         NOTE: This is a *user* warning. Not a developer warning! So we'll use
         the logging module instead of the warnings module!
         """
-        with patch('config_resolver.Config.check_file') as checker_mock:
+        with patch('config_resolver.core.Config.check_file') as checker_mock:
             checker_mock.return_value = (True, "")
-            logger = logging.getLogger('config_resolver')
-            logger.setLevel(logging.DEBUG)
-            catcher = TestableHandler()
-            logger.addHandler(catcher)
             Config('hello', 'world')
             expected_message = (
                 "DEPRECATION WARNING: The file '{home}/.hello/world/app.ini' "
@@ -473,7 +461,7 @@ class FunctionalityTests(unittest.TestCase):
                 "config_resolver! You can already (and should) move the "
                 "file!".format(
                     home=expanduser("~")))
-            catcher.assert_contains(
+            self.catcher.assert_contains(
                 'config_resolver.hello.world',
                 logging.WARNING,
                 expected_message)
@@ -482,13 +470,9 @@ class FunctionalityTests(unittest.TestCase):
         """
         When getting a version number mismatch, the filename should be logged!
         """
-        logger = logging.getLogger('config_resolver')
-        logger.setLevel(logging.DEBUG)
-        catcher = TestableHandler()
-        logger.addHandler(catcher)
         Config('hello', 'world', search_path='testdata/versioned',
                version='2.0')
-        catcher.assert_contains_regex(
+        self.catcher.assert_contains_regex(
             'config_resolver.hello.world',
             logging.WARNING,
             'testdata/versioned/app.ini')
@@ -497,21 +481,18 @@ class FunctionalityTests(unittest.TestCase):
         """
         When getting a version number mismatch, the filename should be logged!
         """
-        logger = logging.getLogger('config_resolver')
-        logger.setLevel(logging.DEBUG)
-        catcher = TestableHandler()
-        logger.addHandler(catcher)
         Config('hello', 'world', search_path='testdata/versioned',
                version='5.0')
-        catcher.assert_contains_regex(
+        self.catcher.assert_contains_regex(
             'config_resolver.hello.world',
             logging.ERROR,
             'testdata/versioned/app.ini')
 
 
-class Regressions(unittest.TestCase):
+class Regressions(TestBase):
 
     def setUp(self):
+        super(Regressions, self).setUp()
         self.cfg = Config('hello', 'world', search_path='testdata')
 
     def test_multiple_log_prefixes(self):
@@ -523,6 +504,64 @@ class Regressions(unittest.TestCase):
         cfg = Config('foo', 'bar')
         self.assertEqual(len(cfg._log.filters), 1)
 
+
+class ConfigResolver5Transition(TestBase):
+    '''
+    To make upgrading to 5.0 easier, we will add a transition layer. This
+    test-case tests that the transition layer calls the old constructor
+    properly and emit appropriate deprecation warnings.
+
+    We can implement this easily as the main entry-point changes from
+    ``config_resolver.Config`` to ``config_resolver.get_config``.
+    '''
+
+    def test_ignore_handler(self):
+        '''
+        Version 5 will accept a new "handler" argument. This should be
+        accepted, but ignored in version 4.
+        '''
+        with patch('config_resolver.core.Config') as mck:
+            get_config('foo', 'bar', {}, handler='dummy_handler')
+            get_config('foo', 'bar', {}, 'dummy_handler')
+        mck.assert_called_with('bar', 'foo', filename='config.ini',
+                               search_path=None)
+
+    def test_search_path(self):
+        '''
+        ``search_path`` should be taken from ``lookup_options``
+        '''
+        with patch('config_resolver.core.Config') as mck:
+            get_config('world', 'hello', lookup_options={
+                'search_path': 'testdata:testdata/a:testdata/b'
+            })
+        mck.assert_called_with(
+            'hello', 'world',
+            filename='config.ini',
+            search_path='testdata:testdata/a:testdata/b')
+
+    def test_filename(self):
+        '''
+        ``filename`` should be taken from ``lookup_options``
+        '''
+        with patch('config_resolver.core.Config') as mck:
+            cfg_b = get_config('world', 'hello', lookup_options={
+                'filename': 'test.ini'
+            })
+        mck.assert_called_with('hello', 'world',
+            filename='test.ini',
+            search_path=None)
+
+    def test_new_default_filename(self):
+        '''
+        In config_resolver 5 we will switch to "config.ini" from "app.ini".
+
+        We want the transition-layer to continue working as usual
+        '''
+        with patch('config_resolver.core.Config') as mck:
+            cfg_b = get_config('world', 'hello')
+        mck.assert_called_with('hello', 'world',
+            filename='config.ini',
+            search_path=None)
 
 if __name__ == '__main__':
     unittest.main()