+++ doc/changelog.rst
@@ -1,3 +1,72 @@
+Release 4.2.5
+-------------
+
+Fixes
+~~~~~
+
+* Change from a module-only distrbution to a package (for PEP-561)
+* Make package PEP-561 compliant
+* Add transition function ``config_resolver.get_config`` for a smoother upgrade
+  to v5.0 in the future.
+* Add deprecation warnings with details on how to change the code for a smooth
+  transition to v5.0
+
+
+Release 4.2.4
+-------------
+
+Fixes
+~~~~~
+
+* Improve code quality.
+* Improve log message for invalid config version numbers.
+
+
+Release 4.2.3
+-------------
+
+Fixes
+~~~~~
+
+* Unit tests fixed
+* Added missing LICENSE file
+* Log messages will now show the complete version string
+* Auto-detect version number if none is specifiec in the ``[meta]`` section.
+* Fix travis CI pipeline
+
+
+Release 4.2.2
+-------------
+
+Fixes
+~~~~~
+
+* Python 2/3 class-inheritance fixed.
+
+
+Release 4.2.1
+-------------
+
+Fixes
+~~~~~
+
+* Log message prefixes no longer added multiple times
+
+
+Release 4.2.0
+-------------
+
+Features added
+~~~~~~~~~~~~~~
+
+* Application & Group name is added to log records
+
+Fixes
+~~~~~
+
+* Python 2/3 Unicode fix in log records
+
+
 Release 4.1.0
 -------------